Refactor LinksController and LinksHelper
Description: This commit refactors the LinksController and LinksHelper files. It updates the before_action callbacks to use the %i syntax for multiple actions. It also adds a new method, filtered_visits, to the LinksHelper module for filtering visits based on IP address and date range. Additionally, it adds a form in the _visits_table.html.erb partial to allow filtering visits by IP address and date range.